moved: helpers into claim-reconciler
changed: fixed tests
