+++ pkg/provisioner/reconciler/claim-reconciler/client.go
@@ -1,14 +1,13 @@
-package reconciler_internal
+package reconciler
 
 import (
 	"context"
-
 	"k8s.io/apimachinery/pkg/runtime"
-
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-type InternalClient struct {
+// internalClient wraps related reconciler fields for convenient passing to helpers
+type internalClient struct {
 	Ctx    context.Context
 	Client client.Client
 	Scheme *runtime.Scheme
