+++ pkg/provisioner/reconciler/claim-reconciler/reconiler.go
@@ -16,11 +16,10 @@ import (
 	"github.com/yard-turkey/lib-bucket-provisioner/pkg/apis/objectbucket.io/v1alpha1"
 	"github.com/yard-turkey/lib-bucket-provisioner/pkg/provisioner/api"
 	pErr "github.com/yard-turkey/lib-bucket-provisioner/pkg/provisioner/api/errors"
-	internal "github.com/yard-turkey/lib-bucket-provisioner/pkg/provisioner/reconciler/reconciler-internal"
 )
 
 type objectBucketClaimReconciler struct {
-	*internal.InternalClient
+	*internalClient
 
 	provisionerName string
 	provisioner     api.Provisioner
@@ -40,17 +39,17 @@ func NewObjectBucketClaimReconciler(client client.Client, scheme *runtime.Scheme
 
 	log.Info("constructing new reconciler", "provisioner", name)
 
-	if options.RetryInterval < internal.DefaultRetryBaseInterval {
-		options.RetryInterval = internal.DefaultRetryBaseInterval
+	if options.RetryInterval < DefaultRetryBaseInterval {
+		options.RetryInterval = DefaultRetryBaseInterval
 	}
 	logD.Info("retry loop setting", "RetryBaseInterval", options.RetryInterval)
-	if options.RetryTimeout < internal.DefaultRetryTimeout {
-		options.RetryTimeout = internal.DefaultRetryTimeout
+	if options.RetryTimeout < DefaultRetryTimeout {
+		options.RetryTimeout = DefaultRetryTimeout
 	}
 	logD.Info("retry loop setting", "RetryTimeout", options.RetryTimeout)
 
 	return &objectBucketClaimReconciler{
-		InternalClient: &internal.InternalClient{
+		internalClient: &internalClient{
 			Ctx:    context.Background(),
 			Client: client,
 			Scheme: scheme,
@@ -73,7 +72,7 @@ func (r *objectBucketClaimReconciler) Reconcile(request reconcile.Request) (reco
 
 	var done = reconcile.Result{Requeue: false}
 
-	obc, err := r.claimForKey(request.NamespacedName)
+	obc, err := claimForKey(request.NamespacedName, r.internalClient)
 
 	/**************************
 	 Delete Bucket
@@ -95,11 +94,11 @@ func (r *objectBucketClaimReconciler) Reconcile(request reconcile.Request) (reco
 	/**************************
 	 Provision Bucket
 	***************************/
-	if !r.shouldProvision(obc) {
+	if !shouldProvision(obc) {
 		log.Info("skipping provision")
 		return done, nil
 	}
-	class, err := internal.StorageClassForClaim(obc, r.InternalClient)
+	class, err := storageClassForClaim(obc, r.internalClient)
 	if err != nil {
 		return done, err
 	}
@@ -126,7 +125,7 @@ func (r *objectBucketClaimReconciler) handleProvisionClaim(key client.ObjectKey,
 		err       error
 	)
 
-	obc, err = r.claimForKey(key)
+	obc, err = claimForKey(key, r.internalClient)
 	if err != nil {
 		if errors.IsNotFound(err) {
 			return fmt.Errorf("OBC was lost before we could provision: %v", err)
@@ -149,17 +148,17 @@ func (r *objectBucketClaimReconciler) handleProvisionClaim(key client.ObjectKey,
 		}
 	}()
 
-	bucketName, err := internal.ComposeBucketName(obc)
+	bucketName, err := composeBucketName(obc)
 	if err != nil {
 		return fmt.Errorf("error composing bucket name: %v", err)
 	}
 
-	class, err := internal.StorageClassForClaim(obc, r.InternalClient)
+	class, err := storageClassForClaim(obc, r.internalClient)
 	if err != nil {
 		return err
 	}
 
-	if !r.shouldProvision(obc) {
+	if !shouldProvision(obc) {
 		return nil
 	}
 
@@ -179,24 +178,24 @@ func (r *objectBucketClaimReconciler) handleProvisionClaim(key client.ObjectKey,
 		return fmt.Errorf("provisioner returned nil/empty object bucket")
 	}
 
-	internal.SetObjectBucketName(ob, key)
+	setObjectBucketName(ob, key)
 	ob.Spec.StorageClassName = obc.Spec.StorageClassName
 
-	if ob, err = internal.CreateObjectBucket(ob, r.Client, r.retryInterval, r.retryTimeout); err != nil {
+	if ob, err = createObjectBucket(ob, r.Client, r.retryInterval, r.retryTimeout); err != nil {
 		return err
 	}
 
-	if secret, err = internal.CreateSecret(obc, ob.Spec.Authentication, r.Client, r.retryInterval, r.retryTimeout); err != nil {
+	if secret, err = createSecret(obc, ob.Spec.Authentication, r.Client, r.retryInterval, r.retryTimeout); err != nil {
 		return err
 	}
 
-	if configMap, err = internal.CreateConfigMap(obc, ob.Spec.Endpoint, r.Client, r.retryInterval, r.retryTimeout); err != nil {
+	if configMap, err = createConfigMap(obc, ob.Spec.Endpoint, r.Client, r.retryInterval, r.retryTimeout); err != nil {
 		return err
 	}
 
 	obc.Spec.ObjectBucketName = ob.Name
 	obc.Spec.BucketName = bucketName
-	if err = internal.UpdateClaim(obc, r.InternalClient); err != nil {
+	if err = updateClaim(obc, r.internalClient); err != nil {
 		return err
 	}
 	log.Info("provisioning succeeded")
@@ -207,28 +206,24 @@ func (r *objectBucketClaimReconciler) handleDeleteClaim(key client.ObjectKey) er
 
 	// TODO each delete should retry a few times to mitigate intermittent errors
 
-	cm := &corev1.ConfigMap{}
-	if err := r.Client.Get(r.Ctx, key, cm); err == nil {
-		err = internal.DeleteConfigMap(cm, r.InternalClient)
+	cm, err := configMapForClaimKey(key, r.internalClient)
+	if err == nil {
+		err = deleteConfigMap(cm, r.internalClient)
 		if err != nil {
 			return err
 		}
-	} else if errors.IsNotFound(err) {
-		log.Error(err, "configMap not found, assuming it was already deleted")
 	} else {
-		log.Error(err, "error getting configMap for deletion")
+		log.Error(err, "could not get configMap")
 	}
 
-	secret := &corev1.Secret{}
-	if err := r.Client.Get(r.Ctx, key, secret); err == nil {
-		err = internal.DeleteSecret(secret, r.InternalClient)
+	secret, err := secretForClaimKey(key, r.internalClient)
+	if err == nil {
+		err = deleteSecret(secret, r.internalClient)
 		if err != nil {
 			return err
 		}
-	} else if errors.IsNotFound(err) {
-		log.Error(err, "secret not found, assuming it was already deleted ")
 	} else {
-		log.Error(err, "error getting secret for deletion")
+		log.Error(err, "could not get secret")
 	}
 
 	ob, err := r.objectBucketForClaimKey(key)
@@ -249,7 +244,7 @@ func (r *objectBucketClaimReconciler) handleDeleteClaim(key client.ObjectKey) er
 		return fmt.Errorf("error deprovisioning bucket %v", err)
 	}
 
-	if err = internal.DeleteObjectBucket(ob, r.InternalClient); err != nil {
+	if err = deleteObjectBucket(ob, r.internalClient); err != nil {
 		if errors.IsNotFound(err) {
 			log.Error(err, "ObjectBucket vanished during deprovisioning, assuming deletion complete")
 		} else {
@@ -259,55 +254,21 @@ func (r *objectBucketClaimReconciler) handleDeleteClaim(key client.ObjectKey) er
 	return nil
 }
 
-// shouldProvision is a simplistic check on whether this obc is a concern for this provisioner.
-// Down the road, this will perform a broader set of checks.
-func (r *objectBucketClaimReconciler) shouldProvision(obc *v1alpha1.ObjectBucketClaim) bool {
-	logD.Info("validating claim for provisioning")
-	if obc.Spec.ObjectBucketName != "" {
-		log.Info("provisioning already completed", "ObjectBucket", obc.Spec.ObjectBucketName)
-		return false
-	}
-	if obc.Spec.StorageClassName == "" {
-		log.Info("OBC did not provide a storage class, cannot provision")
-		return false
-	}
-	return true
-}
-
 func (r *objectBucketClaimReconciler) supportedProvisioner(provisioner string) bool {
 	return provisioner == r.provisionerName
-}
-
-func (r *objectBucketClaimReconciler) claimForKey(key client.ObjectKey) (*v1alpha1.ObjectBucketClaim, error) {
-	logD.Info("getting claim for key")
-	obc := &v1alpha1.ObjectBucketClaim{}
-	if err := r.Client.Get(r.Ctx, key, obc); err != nil {
-		if errors.IsNotFound(err) {
-			return nil, err
-		}
-		return nil, fmt.Errorf("error getting claim: %v", err)
-	}
-	return obc.DeepCopy(), nil
 }
 
 func (r *objectBucketClaimReconciler) objectBucketForClaimKey(key client.ObjectKey) (*v1alpha1.ObjectBucket, error) {
 	logD.Info("getting objectBucket for key", "key", key)
 	ob := &v1alpha1.ObjectBucket{}
 	obKey := client.ObjectKey{
-		Name: fmt.Sprintf(internal.ObjectBucketNameFormat, key.Namespace, key.Name),
+		Name: fmt.Sprintf(ObjectBucketNameFormat, key.Namespace, key.Name),
 	}
 	err := r.Client.Get(r.Ctx, obKey, ob)
 	if err != nil {
 		return nil, fmt.Errorf("error listing object buckets: %v", err)
 	}
 	return ob, nil
-}
-
-func (r *objectBucketClaimReconciler) configMapForClaimKey(key client.ObjectKey) (*corev1.ConfigMap, error) {
-	logD.Info("getting configMap for key", "key", key)
-	var cm *corev1.ConfigMap
-	err := r.Client.Get(r.Ctx, key, cm)
-	return cm, err
 }
 
 func (r *objectBucketClaimReconciler) updateObjectBucketClaimPhase(obc *v1alpha1.ObjectBucketClaim, phase v1alpha1.ObjectBucketClaimStatusPhase) (*v1alpha1.ObjectBucketClaim, error) {
@@ -320,13 +281,13 @@ func (r *objectBucketClaimReconciler) updateObjectBucketClaimPhase(obc *v1alpha1
 }
 
 func (r *objectBucketClaimReconciler) deleteResources(ob *v1alpha1.ObjectBucket, cm *corev1.ConfigMap, s *corev1.Secret) {
-	if err := internal.DeleteObjectBucket(ob, r.InternalClient); err != nil {
+	if err := deleteObjectBucket(ob, r.internalClient); err != nil {
 		log.Error(err, "error deleting objectBucket")
 	}
-	if err := internal.DeleteSecret(s, r.InternalClient); err != nil {
+	if err := deleteSecret(s, r.internalClient); err != nil {
 		log.Error(err, "error deleting secret")
 	}
-	if err := internal.DeleteConfigMap(cm, r.InternalClient); err != nil {
+	if err := deleteConfigMap(cm, r.internalClient); err != nil {
 		log.Error(err, "error deleting configMap")
 	}
 }
